Initial commit
This commit contains a simple README.md file elaborating what the
content of this repository is, as well as the MIT license file.

It also adds a simple template for creating tasks as well as the first
two tasks, pulling a docker image as well as creating an ubuntu based
container, running 'uname -s' inside it and finally deleting the stopped
container.