Finish instruction and solution for task 3
This commit creates the instructions and solution for task 3, in which
the student is supposed to create a new container which publishes a
webpage to port 80.

The student is then required to configure the container network to
forward port 80 to the host system.

As a second task, the containers publishing port can be reconfigured by
supplying the env var PORT to the container. The student is therefore
required to configure the container to publish to port 8080 while
forwarding the newly configured port to port 80 on the host system.